backend/requirements.txt
fastapi
uvicorn
sqlalchemy
pydantic
python-dotenv
google-api-python-client
google-auth
slowapi
